Refactorizations: - Store time as float, resolving several compilation issues with clang - Added missing virtual destructor in GUI - Fixed typo in OS_WIN32_MINGW (bh.h)

+++ branches/SFML2/src/config_screen.cpp
@@ -171,7 +171,7 @@ void CGameConfig::Enter() {
 	Music.Play(param.config_music, true);
 }
 
-void CGameConfig::Loop(double time_step) {
+void CGameConfig::Loop(float time_step) {
 	ScopedRenderMode rm(GUI);
 	Winsys.clear();
