+++ branches/SFML2/src/newplayer.h
@@ -23,7 +23,7 @@ GNU General Public License for more details.
 
 class CNewPlayer : public State {
 	void Enter();
-	void Loop(double time_step);
+	void Loop(float time_step);
 	void Keyb(sf::Keyboard::Key key, bool special, bool release, int x, int y);
 	void TextEntered(char text);
 	void Mouse(int button, int state, int x, int y);
